cos/cross_binding.rs: colocate 17 tests with production (#984 P3 Phase 2b)
Move 17 tests targeting redirect_local_cos_request_to_owner,
prepared_cos_request_stays_on_current_tx_binding, and
resolve_local_routing_decision from tx/mod.rs::tests to a new
mod tests in cos/cross_binding.rs.

New tests block adds explicit imports for PROTO_TCP,
COS_MIN_BURST_BYTES, and SharedCoSQueueLease (production scope of
cross_binding.rs doesn't glob these).

Mechanical: 17 tests, 531 lines added; 515 removed from tx/mod.rs.
cargo test --bins 865/0/2.
